auth-service/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.28.1
